@@ -18,7 +18,7 @@
         "eslint": "8.39.0",
         "eslint-config-next": "13.3.4",
         "framer-motion": "^10.12.8",
-        "next": "13.3.4",
+        "next": "^14.2.5",
         "next-intl": "^3.15.3",
         "postcss": "8.4.23",
         "react": "18.2.0",
