@@ -1,3 +1,7 @@
+import createNextIntlPlugin from 'next-intl/plugin';
+
+const withNextIntl = createNextIntlPlugin();
+
 /** @type {import('next').NextConfig} */
 const nextConfig = {
 	reactStrictMode: true,
@@ -14,10 +18,6 @@ const nextConfig = {
 			},
 		],
 	},
-	i18n: {
-		locales: ['en', 'fr'],
-		defaultLocale: 'en',
-	},
 };
 
-module.exports = nextConfig;
+export default withNextIntl(nextConfig);
