@@ -1,3 +1,5 @@
+'use client';
+
 import { useState } from 'react';
 import { PhotoAlbum } from 'react-photo-album';
 import styled from 'styled-components';
@@ -12,56 +14,56 @@ import { Link } from '@/styles/common';
 import NextJsImage from './NextJsImage';
 
 interface Props {
-    data: Image[];
-    category: 'people' | 'places';
+	data: Image[];
+	category: 'people' | 'places';
 }
 
 const Photos = ({ data, category }: Props) => {
-    const [index, setIndex] = useState(-1);
-    const scrollToTop = () => window.scrollTo({ top: 0, behavior: 'smooth' });
+	const [index, setIndex] = useState(-1);
+	const scrollToTop = () => window.scrollTo({ top: 0, behavior: 'smooth' });
 
-    return (
-        <>
-            <PhotoAlbumWrapper>
-                <PhotoAlbum
-                    photos={data}
-                    layout={'masonry'}
-                    renderPhoto={NextJsImage}
-                    columns={(containerWidth) => {
-                        if (containerWidth < 500) return 1;
-                        if (containerWidth < 1000) return 2;
-                        return 3;
-                    }}
-                    onClick={({ index }) => setIndex(index)}
-                />
-            </PhotoAlbumWrapper>
-            <Lightbox
-                slides={data}
-                open={index >= 0}
-                index={index}
-                close={() => setIndex(-1)}
-                plugins={[Thumbnails, Zoom]}
-            />
-            <LinkWrapper>
-                <Link href='#' onClick={() => scrollToTop()}>
-                    Back to Top
-                </Link>
-                <Link href={`/${category === 'people' ? 'places' : 'people'}`}>
-                    {category === 'people' ? 'Places →' : 'People →'}
-                </Link>
-            </LinkWrapper>
-        </>
-    );
+	return (
+		<>
+			<PhotoAlbumWrapper>
+				<PhotoAlbum
+					photos={data}
+					layout={'masonry'}
+					renderPhoto={NextJsImage}
+					columns={(containerWidth) => {
+						if (containerWidth < 500) return 1;
+						if (containerWidth < 1000) return 2;
+						return 3;
+					}}
+					onClick={({ index }) => setIndex(index)}
+				/>
+			</PhotoAlbumWrapper>
+			<Lightbox
+				slides={data}
+				open={index >= 0}
+				index={index}
+				close={() => setIndex(-1)}
+				plugins={[Thumbnails, Zoom]}
+			/>
+			<LinkWrapper>
+				<Link href='#' onClick={() => scrollToTop()}>
+					Back to Top
+				</Link>
+				<Link href={`/${category === 'people' ? 'places' : 'people'}`}>
+					{category === 'people' ? 'Places →' : 'People →'}
+				</Link>
+			</LinkWrapper>
+		</>
+	);
 };
 
 export default Photos;
 
 const PhotoAlbumWrapper = styled.div`
-    padding-bottom: 15rem;
+	padding-bottom: 15rem;
 `;
 
 const LinkWrapper = styled.div`
-    display: flex;
-    justify-content: space-between;
-    padding-bottom: 3rem;
+	display: flex;
+	justify-content: space-between;
+	padding-bottom: 3rem;
 `;
