@@ -1,4 +1,5 @@
-import { GetStaticProps } from 'next';
+'use client';
+
 import { useTranslations } from 'next-intl';
 
 import Layout from '@/components/Layout';
@@ -7,9 +8,9 @@ import Photos from '@/components/Photos';
 import useGetPhotos from '@/hooks/useGetPhotos';
 import { Heading, HeadingWrapper, Link, SubHeading } from '@/styles/common';
 
-const People = () => {
-	const { data, loading } = useGetPhotos('people');
-	const t = useTranslations('People');
+const Places = () => {
+	const { data, loading } = useGetPhotos('places');
+	const t = useTranslations('Places');
 
 	return (
 		<Layout>
@@ -18,19 +19,11 @@ const People = () => {
 					<Heading>{t('heading')}</Heading>
 					<SubHeading>{t('subheading')}</SubHeading>
 				</div>
-				<Link href='/places'>{t('placesLink')}</Link>
+				<Link href='/people'>{t('peopleLink')}</Link>
 			</HeadingWrapper>
-			{loading ? <LoadingIndicator /> : data && <Photos data={data} category={'people'} />}
+			{loading ? <LoadingIndicator /> : data && <Photos data={data} category={'places'} />}
 		</Layout>
 	);
 };
 
-export default People;
-
-export const getStaticProps = (async (context) => {
-	return {
-		props: {
-			messages: (await import(`public/locales/${context.locale}.json`)).default,
-		},
-	};
-}) satisfies GetStaticProps;
+export default Places;
