@@ -1,5 +1,5 @@
 import createMiddleware from 'next-intl/middleware';
-import { locales } from './config';
+import { locales } from '../config';
 
 export default createMiddleware({
 	// A list of all locales that are supported
